chore: Refactor according to coding guidelines (#3399)
