+++ src/api/vipps-login/api.ts
@@ -1,4 +1,4 @@
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 import client from '@atb/api/client';
 import {VippsSignInErrorCode} from '@atb/auth/AuthContext';
 import {generateNonce, generateState} from '@atb/api/vipps-login/utils';
