+++ src/api/vipps-login/utils.ts
@@ -1,5 +1,5 @@
 import {v4 as uuid} from 'uuid';
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 
 export const generateState = async () => {
   const state = uuid();
