+++ src/chat/user.ts
@@ -1,7 +1,7 @@
 import Intercom from 'react-native-intercom';
 import {Dimensions, PixelRatio, Platform} from 'react-native';
 import DeviceInfo from 'react-native-device-info';
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 import {checkGeolocationPermission} from '@atb/GeolocationContext';
 import {updateMetadata} from './metadata';
