+++ src/components/checkbox/index.ts
@@ -1,0 +1,1 @@
+export {Checkbox} from './Checkbox';
