+++ src/debug/use-debug-override.ts
@@ -1,5 +1,5 @@
 import {useEffect, useState} from 'react';
-import storage, {StorageModelKeysEnum} from '@atb/storage';
+import {storage, StorageModelKeysEnum} from '@atb/storage';
 import {parseBoolean} from '@atb/utils/parse-boolean';
 
 export type UseDebugOverride = [boolean | undefined, (v?: boolean) => void];
