+++ src/global-messages/storage.ts
@@ -1,4 +1,4 @@
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 import {GlobalMessageType} from '@atb/global-messages';
 
 export const DISMISSED_GLOBAL_MESSAGES_KEY =
