+++ src/index.tsx
@@ -9,7 +9,7 @@ import {RootStack} from '@atb/stacks-hierarchy';
 import trackAppState from './diagnostics/trackAppState';
 import ThemeContextProvider from './theme/ThemeContext';
 import {FavoritesContextProvider} from './favorites';
-import SearchHistoryContextProvider from './search-history';
+import {SearchHistoryContextProvider} from './search-history';
 import {TicketingContextProvider} from './ticketing';
 import RemoteConfigContextProvider from './RemoteConfigContext';
 import {AuthContextProvider} from './auth';
