+++ src/login/LoginOptionsScreen.tsx
@@ -12,7 +12,7 @@ import * as Sections from '@atb/components/sections';
 import {ThemeText} from '@atb/components/text';
 import {VippsLoginButton} from '@atb/components/vipps-login-button';
 import {AfterLoginParams, LoginInAppScreenProps} from '@atb/login/types';
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 import {StyleSheet} from '@atb/theme';
 import {StaticColorByType} from '@atb/theme/colors';
 import {LoginTexts, useTranslation} from '@atb/translations';
