+++ src/search-history/index.ts
@@ -1,0 +1,5 @@
+export {
+  SearchHistoryContextProvider,
+  useSearchHistory,
+} from './SearchHistoryContext';
+export type {JourneySearchHistoryEntry} from './SearchHistoryContext';
