+++ src/stacks-hierarchy/Root_MobileTokenOnboarding/components/InspectableTokenInfo.tsx
@@ -7,7 +7,7 @@ import {StyleSheet} from '@atb/theme';
 import React from 'react';
 import {StaticColorByType} from '@atb/theme/colors';
 import useFocusOnLoad from '@atb/utils/use-focus-on-load';
-import TravelTokenBox from '@atb/travel-token-box';
+import {TravelTokenBox} from '@atb/travel-token-box';
 import {RemoteToken} from '@atb/mobile-token/types';
 import {isTravelCardToken} from '@atb/mobile-token/utils';
 import {useAppState} from '@atb/AppContext';
