+++ src/stacks-hierarchy/Root_PurchaseConfirmationScreen/components/SelectPaymentMethodSheet.tsx
@@ -26,7 +26,7 @@ import LinearGradient from 'react-native-linear-gradient';
 import PaymentBrand from './PaymentBrand';
 import {useFirestoreConfiguration} from '@atb/configuration/FirestoreConfigurationContext';
 import {getExpireDate, getPaymentTypeName} from '../../utils';
-import {Checkbox} from '@atb/components/checkbox/Checkbox';
+import {Checkbox} from '@atb/components/checkbox';
 
 type Props = {
   onSelect: (value: PaymentMethod) => void;
