+++ src/stacks-hierarchy/saved-payment-utils.ts
@@ -4,7 +4,7 @@ import {
   RecurringPaymentWithoutCardOption,
   SavedPaymentOption,
 } from './types';
-import storage from '@atb/storage';
+import {storage} from '@atb/storage';
 import Bugsnag from '@bugsnag/react-native';
 import {FirebaseAuthTypes} from '@react-native-firebase/auth';
 import {listRecurringPayments} from '@atb/ticketing';
