+++ src/travel-token-box/index.ts
@@ -1,0 +1,1 @@
+export {TravelTokenBox, TravelDeviceTitle} from './TravelTokenBox';
